requirements.txt
# ── API & DB ──────────────────────────────────────
fastapi==0.111.0
uvicorn[standard]==0.29.0
asyncpg==0.29.0
psycopg2-binary==2.9.9      # used only by seed_data.py

# ── Data / Seed ───────────────────────────────────
faker==24.11.0

# ── Dev / Test ────────────────────────────────────
pytest==8.2.0
httpx==0.27.0               # async test client for FastAPI
pytest-asyncio==0.23.7
